DATABASE UPDATE * Added new production line and building: liquor -> get-together at tavern   (placeholder graphics and icons, build menu xml, DB)
Note that currently SettlerServiceProviders have no collector unit, thus the tavern won't get any liquor. This needs a fix
as well as the settler behaviour -- I managed to add the get-together consumption line but the SettlerCollector does not
collect stuff from tavern even if it procudes get-together.
The name 'get-together' is not fixed in any way and might be replaced by one looking better ingame.