* Move current pot template
After an update as described in po/README, the generated file will have the name unknown-horizons.pot as well, so I rename this.
Still need to do the renaming internally in pootle although one can upload files 'To:' a differently named one -- which I propose as viable solution until this transition is made.

Some of my recent commits hopefully fixed the weird issue with non-translated tooltips.