* Display proper names in the inhabitant overview tab instead of 'Tent'
This introduces a new column residential_name in settler/settler_level.
I as well added three more increments with data to the database. By no means decided on even the amount of increments we plan to have, just wanted to remind that we at some point need to start this discussion.