Be even more paranoid about the environment of game tests
setup/teardown of function based tests can't pass arguments to the test, or
keep a reference somewhere. If a test fails, we need to reset the session under
all circumstances, otherwise we break the rest of the tests. The global
database reference has to be set each time too. Unit tests also use it,
and we can't predict the order tests run (we should not rely on it anyway).

While we're at it, we cut down on some boilerplate and let the decorator
handle the creation of a new session. We could use this later to define
different preconditions for a test, e.g. only a new session, or a
prebuild settlement, different map etc.