* Initial work on new voice file structure
Plan: Game tells us which event triggered, we check that folder for a folder of the currently set locale.
If such a folder is found (example files: de/), a random file from within that folder is played.
If no locale is found, we use a random file from en/ instead; we do not play any speech (and perhaps print a warning to shell) should that not exist as well.