* Don't translate savegame timestamp in GUI
Localizing this did not work as intended (at least I guess it was).
Should a feature request for custom date layout reach us, we will need
to find a better solution, wrapping this into a string renders it
completely useless because translators at pootle have a hard time
figuring out the syntax of date / strftime.