* Modify inhabitant happiness formula for taxes once more
Now, at tax rate 1.0, inhabitant happiness will
* very slightly decrease for sailors
* preserve the status quo for pioneers
* slightly increase for settlers
All of the above holds if players manage to deliver most goods that are
desired in the respective increment.
Because there are more of these resources for settlers, I decided to
give them a (rather small) leg up in comparison to sailors and pioneers.