* Use clickable labels to select color instead of ListBox strings #1364
This adds a new setting ColorID to the settings file. Partially fixes a
bug caused by the different SP menu modes. Still observed when trying to
enter a new player name and then clicking on a different SP game mode.
The text in the player name input field is reset during this process.