Fix IngameType class_package overriding
When loading objects the class_package and class_name attribute were
overwritten by each new building/unit. These attributes have to be set
on the created class, not the metaclass.
This created an issue with the BuildingTool, where it showed the info
message "you can drag to build roads" for every object type, not just
paths.