Add integrated IPython shell for development
Starts an IPython kernel, that's basically an interactive shell running
in uh process. However the input is handled by frontends running in other
processes. Once this kernel is running, you can connect to it with:

  ipython console --existing

This allows you to influence the game (work with the session, add
buildings etc.). Therefore it works similar to the existing in-game
console, but is way more comfortable.

Needs IPython 0.12.