* Fix highly weird crash with unicode game messages
To reproduce the initial crash, set your locale to German and start a
game, then open the logbook on the chat page and try to send a message.
The update triggered the display of the new world message, which
contains umlauts in its German translation.
Trying to display them in the listbox resulted in an UnicodeDecodeError.

Still need to investigate why calling unicode() on the message (just
like done with the chat messages) would not suffice in this case.