Fix label update of current players in MP menu
When registering _cb_game_details_changed for "lobbygame_changename"
callback there was the problem that this callback would call functions
with 5 arguments but _cb_game_details_changed understood only 3.

To circumvent future problems of this type, one could use messaging
instead or check if the functions registered as callbacks understand
the correct number of arguments.