Fixed segfault when pressing F9 Thanks @ ChrisOelmueller
The problem is twofold:
On KeyEvents, the keylisteners are called one after another.
The first one would trigger quickload, which causes other listeners to be removed
and new ones to be added. Those changes will lazily take effect just on the next event.
Therefore, the loop that handles the first F9 will gladly continue its iteration
and call the next listener, which is dead already but not removed as it just happens on the next event.

This is now prevented by consuming the event in the first listener,
in order to make the handling loop not call any other listeners.
Here, it's even the desired behaviour, but in general, it might not be.