Try to detect invalid events from pychan.
Sometimes, pychan sends invalid event notifications. It is probably related to changing the gui when the mouse hovers on gui elements. Random tests have given evidence to believe that pychan indicates invalid events by setting the top container's position to 0, 0. Since this position is currently unused, it can serve as invalid flag, and dropping these events seems to lead to the desired placements.

A proper fix would probably be related to this:
http://fife.trac.cvsdude.com/engine/ticket/387

Refs #1717