* Changes which crash right now, but in my opinion should work
With the prints committed, the build preview uses the current
inhabitant level as parameter to look up the preview action set.
Upon trying to place the actual building, the action set used for
preview is passed to the function again, but now uses the default
level=0. For any building that does not have a base action set in
TIER.SAILORS (0), the preview will work but placing the building won't.