fix crash an leaving random map menu too early
If the random map menu is left before the random map update took
place, there was a crash because the needed fields could not be found
any more.
Try this: in the random map selection menu change a slider and press
ESC immediatly afterwards. Then wait a litte in the main menu. The
game should crash.