Don't crash if no persistant cache file can be opened
If no persistant cache file can be opened, the game retries and then
crashes. Imho the game should still be playable even if the startup
takes a little longer - at least it is playable at all.
On the down side we might not get as many crashes reported and a real
bug might get hidden. Still I think it's worth it.