* Implement skeleton for #1671 Balance display in top left corner
Currently displays either a red or green arrow indicating negative or
positive balance. Toggle display by clicking on the gold icon.

Still needs a proper way of displaying the actual balance (see TODO).
Especially noticeable is the fact that (for very small settlements)
decommissioning buildings does not trigger a display update.
This definitely needs to be fixed since a common use case for this mode
would be decommissioning enough buildings to reach a positive balance.