* Refactored code that plays background music
Old and new behavior:
* When a game is started or loaded, play the main menu tracks
* After this track is finished, play ingame tracks even if in main menu
  since we only have one special menu track right now

Updated behavior:
* Keep a deque of recently played tracks to avoid repetition

Especially not changed was the fact that the first track to be played is
the current main menu background music, however we may start the game.
If you consider this a bug instead of our signature music, we need to
use the newly introduced (and almost unused) refresh_playlist parameter.