* Refactor logging init and remove duplicated code
Fife's LogManager works on bools, we started with bools, but somehow
during the process we work with ints (and store them in the settings).
The settings template was updated to use bools.

Also removes apparently unused definition of 'self.log', instead only
defines 'self._log' for fife, and an assertion which i could not justify.