* Fix pastry shop overview crashing. Still looks bad.
Right now we still display all available lines if there is a storage or
warehouse in range. Since we cannot make further assumptions about
resource availability, the pastry shop most likely needs an own
interface implementation which makes use of the fact that the possible
lines are 1+1 permutations of 2+2 resources.