* Dynamic placement of cityinfo (depends on resbar size)
Adding items to the bar should now shift the cityinfo to the right until
meeting the minimap, and then instead move it to the upper screen edge.

This was suggested by Traveler in our forums. Thanks!

A remaining problem: When hovering the settlement after the cityinfo was
hidden, it will be centered again just like before.
Hopefully we can fix this rather soon, else we might want to revert.