Fixed ticket #1381.
Implemented system in which Message has a type that determines what happens on click.

I was using current_page instead of _cur_entry. Changed current_page to current_mode to clarify what it's used for. That wasn't the only problem, but clicks on message notifications now open their relevant logbook page.

Known bug:
If you move the cursor to the left edge of the screen so that the screen scrolls, then click on a message, the logbook will pop up and the screen will continue scrolling left until you close the logbook.

Removed debug logging and commented-out code.

Message displayed status is now saved to fix a bug where when a game was loaded, all previous messages would be displayed any time a new message was displayed.

Incremented SAVEGAMEREVISION.