* Remove pigsty, introduce herbary
Cattle are one of our main meat sources and the pigsty design did not
really work out in practice, being a larger field than everything else.

It was planned to cut back on food supply chains for a long time now.
We might consider adding more of those lines if the meta-resources
significantly change gameplay again in terms of different food resources.