* New component RestrictedPickup (right now used for SettlerCollector)
Treats something as allowed (not restricted) if it is contained in the
whitelist "allowed". If you wish your collector to have no restrictions,
just do not add this component to its object file.

The presence of RestrictedPickup with an empty "allowed" line will
forbid the collector from picking up anything.