* Routeconfig and buyselltab layout overhaul
Now features red and green icons again. I am aware of the implications,
but still hope that this is an improvement over the previous situation.

While I'd really like to retain the clear statement especially with
regard to the hover functionality, there are problems with
internationalizing the icon and transferring 'buy' to 'load' for the
routeconfig would have been over the top for me.