* Use YAML anchors in tutorial as proof of concept for #1573
Every anchored argument in "logbook-data" (which is an arbitrary and
currently unused name) is also available in the parsed "events" output.

An obvious drawback of this approach is that logbook-data needs to be
defined before the actual scenario logic description can happen.
Since this cannot work with anchors, an implementation that puts the
string database below the scenario logic would most probably require
additional code on our side. For now, this should work though.