* Made boat builder gui dynamic
Now loads the list of ships defined for each tab and automatically
displays the construction cost and images we previously hardcoded
into xml files. Adds a very simple showcase tab instead of those.

Right now the only visible issue is that the gold cost is always zero.
This happens because the current code uses the boat builder Producer
to generate production lines, but the gold cost is not paid from the
boat builder inventory (directly paid by the player). Should fix this.