* Also allow querying for component names as strings
This decision is certainly not easy to defend, but it does prevent
prevent fun with circular imports if you need components in places
like util.uhdbaccessor.
The error detection would also not work as intended with arbitrary
strings passed, so i'd be very happy too see somebody stepping up
and fixing stuff, effectively reverting this commit.