* Make action "increase game speed" only have one assigned hotkey
This effectively reverts fdf07d0e348fd9495270a6ab8c77e24c9d66e955.

With the hotkey design approach I took for now, we do look at actions
as main source of information when configuring the interface, not keys.
Internally we still store this information keyed by the associated key.
This means that while the same action xml-technically may be assigned
to multiple actions, only one of them would be recognized (unordered).

For devices where + and = would not send the same keycode (for me they
do) and the default + is inconvenient (such as US keyboard layouts),
this is an excellent opportunity to try and reconfigure the hotkey :)