* Remove code that is no longer required
As of 0070d652e17caf1a40bf314e1b4eeba7a69aa2cd we always select an item.
If there is no such item, the dialog is interrupted from the main menu
already, which means that something thoroughly went wrong if we still
see -1 as selected item of the list box.