* Interface to query default hotkeys from settings
Also removes the keyname mappings of +-,. -- while a nice idea on paper,
this caused too many issues everywhere to be a reasonable design choice.

Introduces a validation for key input from user settings files.
If we detect invalid keys, this action is reset to the default key.

New feature: Clicking on the line that states the current and default
key will also assign the default key to the current action.