Fix example gui test
The assertion about the active widgets used to be pretty simple and was
used at the beginning of working on the gui test system, but
unfortunately got blown up in complexity.

The test broke because there are now 2 active widgets, the background and the
loading screen.

Removing this should be okay, it didn't serve any real purpose.