Fixes #1833 - Display if a savegame is upgradable.
Note that the code relies on the fact that we have incrementally rising
upgrade methods. (upgrade_to_revX) That means if we have an upgrade that
does not need a upgrade function, we will still have to have a upgrade
function to keep this code working. In return this code does not rely on
any other variables being kept up to date.