Fix #1909: Object translations not updated when changing language
In this case, starting with the System Default language option and
then swapping to English would not reset the object translations
since English is our fallback translation (i.e. represented by a
NullTranslation). Now the object holding all translations that is
supposed to be imported everywhere is updated to account for this.

Thanks @squiddy for suggesting this in a commit comment on #1905!