tests: Improved search of child widgets and test enhancements
Widgets like ScrollArea have children, but no attribute with that name.
In that case, we use findChildren.

Also extend the found widget instances with helper functions for easier
testing. Right now, we extend the ListBox widget with a select method,
that can be used to select a value and trigger callbacks.