Move sql template strings to horizons/i18n/
While they probably do belong into po/ this is more likely to cause
confusion (since the file is not meant to be used as template) --
it is concatenated with the main interface template while still under
version control on its own.
The file also does not contain a proper .pot header.

All in all, since we symlink po/ into pootle, cleaning that folder up
makes sense.