tests: Add support for sliders in tests and gui logger
I haven't found a way to detect the start/end of changing a slider yet.
Instead, we receive many slider-changed events. We will only create code
once events for a new widget are received. The code will use the latest
value of the slider.

This means that in order to see code for a changed slider, you should
trigger and event for a different widget.