Minimap rectangle should have cursor centered, regardless of zoom level (#1969)
When the cursor clicks on an area in the minimap, the rectangle should
indicate the displayed area, with the cursor centered inside it. To get
the cursor centered rather than in the upper left corner, we move this
corner by an offset calculated by the width and height of the rectangle.
However, up until now, this offset did not include zoom level which
means the cursor was not centered properly when zooming out.