Move keypress handling to IngameGui
This is part of an effort to make IngameGui responsible for all parts of
the game's UI. By that it becomes easier to swap out against a different
GUI, e.g. for the editor.

We're still working on trimming down the size of both Gui and IngameGui
by moving code into widgets/windows, once that is done we might be able
to move all of the keylistener code into these classes.