Remove another use of switch_current_widget
Technically, it's still calling the same code, but this way it's easier
for me to see what's happening and to move such functionality into the
window manager.

Also changes some code in the editor start menu to keep show/hide
as methods that will only be called from the outside.