Rewrite singleplayer menu
After probably 10 attempts to improve the code in small steps and not
breaking tests I gave up and rewrote it in one big change.

This splits the different map modes into separate classes. The
SinglePlayerMenu is only responsible to switch between modes.

Note: the map preview is not yet integrated