style: Make 'book' style the default style
Remaining use cases of 'default' were simply mistakes with a conversion
of some sorts. Of course, as with all style-related changes, this commit
is not supposed to change anything but will mangle our gui in practice.

Special handling for the Label and Container/H-/VBox was ported over
from 'book' without investigating further; same for the arbitrary and
inconsistent attribute order throughout the different widgets.