style: Remove references to 'book' style
Now the default style.

Remove style parameter from the pickbelt widget because all widgets used
'book' as style anyways and it seems likely that it will stay that way.

Multiplayer menu was not edited and will continue to map an empty style.
This should mean no harm since a rewrite is scheduled anyways and empty
styles will just use the default one.