style: Drop tuple notation in styles. Refs #1335
Even after attempting to fix it on UH side, it remains buggy and not
ready for production. If at some point in the future this is unbroken
by FIFE folks, we can still combine the few widgets where it matters.

Getting rid of Tooltip* widgets helped to make this a non-issue.