Remove dialog_executed flag
To be honest, I don't understand why showing another dialog will lead to
an endless loop.
Anyway, I'm going to rewrite the multiplayer menu and error handling
will likely change as well, so perhaps this won't be an issue in the
future.

If this problem surfaces again, I think we should find a different
solution. I don't want windows/dialogs to peek into internals of the
manager.