Start implementing the WindowManager
The WindowManager will be used to open new windows, close existing ones.
By keeping track of all open windows, it is easy to restore previous
windows that were hidden when a new window was shown.

In this commit, only sp/mp menu are managed. I'm trying to not break
everything, but it's very likely that bugs are introduced. Those will be
gone once every window is managed.