Move modal background to window class
This way every window can benefit from that feature. Normal windows have
to handle the show/hide of the background themselves, dialogs may set
`modal` to True on the class or instance level.

This temporary removes the modal functionality from the old dialogs, but
they will be ported over to the new dialog window soon.