Fix bug when showing a dialog that was closed before
The bug caused this (wrong) behaviour:

  1. open load game
  2. click delete button
  3. close confirmation popup
  4. close load game dialog
  -> Game exits instead of showing the main menu

The problem was that the instance of the SelectSavegame dialog is reused
after the popup is closed. However, the dialog was closed (thereby
exiting the dialog mainloop). This incorrect flag caused `show` to think
that the dialog was just hidden, and not enter a new dialog mainloop.
Trying to close the dialog breaks out of the current mainloop, which
happens to be the game's mainloop.