Hide minimap widget when ending a game
I just found out about this bug because I was running a higher
resolution. Before the mainmenu's background hid the minimap. :)

Also moved the unsubscription of messages to the top of the `end()`
method, because I noticed a crash when a SpeedChanged message came in
after the minimap instance was gone.