Avoid new random map previews when values are changing rapidly
If the sliders are modified quickly, we are constantly triggering new
random map generations that are killed milliseconds later.

We might have to tweak the value of 0.5s, but I think it's reasonable to
do this (even with a lower time), the small latency on very fast
machines should be negligible.