Fix #2000 Certain ToggleImageButtons segfault on hover
While not explicitly setting those attributes is fine (implicit handling
in FIFE/guichan is to use up_image in this case), we were storing
references to those objects in old_images for ToggleImageButton and thus
creating invalid images when trying to reload those. This fix thus just
makes the handling more transparent both to FIFE and code-wise.