Editor GUI uses its own SelectSavegameDialog instance
An instance of a window is bound to a specific window manager. The old
code used the main_gui's dialog, which means that the ingame window
manager never knew about this new window.

As a side effect, the editor ingame gui doesn't need a reference to the
main gui anymore. Yay.