Switch keys on duplicate.
If the user tries to bind a key which already has a binding, instead of simply removing the old binding, I bind the key binded to the current action to the other action. This is done to prevent having an action without a corresponding key, which is currently not supported by the hotkey implementation.