Remove unused attribute in Dialog
Because it looked like a spelling mistake, attempts were made to rename
the attribute to return_events (without leading underscore).

This however broke a lot of code, because popups used to set the return
events as class attributes. 'self.return_events = {}' in the class'
__init__ shadowed the class attribute which led to an empty dict.

The recent change that moved the popup code into a Dialog subclass
removed this problem (because return_events are set up in prepare),
removing this obsolete attribute is still the correct thing to do.