Island fertility
Yes, this is gathering resource<->building relations in code instead of
our database. Sorry about that. Will look into wiring those in later.

Currently assumes a 0/1 fertility system (either resource A can exist on
island, or it cannot). There are no stepdowns like "half efficiency" for
now and i'm not sure they add that much gameplay value, as opposed to
interface burden and code complexity (which they definitely do add).

As you may have noticed this commit only displays some random icons for
settlements. I'm looking for an implementation of the underlying concept,
even if it's rough and unpolished.
The current mapping for resources against buildings indicates that if an
island is infertile w.r.t. a particular resource, the corresponding
building in MAPPING should be impossible to build in settlements on that
island. Whether to indicate that from our build menus, message widget,
buildability highlights, or a combination thereof, is your call.