Script to translate a scenario
Based on removed second half of create_scenario_pot.

This script now really only creates a scenario .pot and leaves all the
dirty work to `development/translate_scenario.py`.
Change was necessary because i still have not given up on automating all
of this through pootle, and python hooks integrated significantly better
with pootle than bash ones in my test. In particular, maintaining a
python script embedded in bash is not what i was looking for, and
hopefully we'll convert the remainining python-bash mixture as well.

Known regressions: Metadata currently not overwritten
Because of that, scenarios were not updated yet.