overlays: Start work on inventory overlay component
Adds an initial structure for the component, with somewhat well-defined
methods and a few properties as more readable shortcuts for FIFE-based
stuff that is needed every now and then.

Also adds a bit of documentation, some TODOs and error handling for the
yaml-defined component data (no checks for negative amounts currently).

As for `remove()`:
I left some questions about how the base image is supposed to be handled
in the respective code. This should ideally be clarified with FIFE to not
cook up some custom haphazard "solution" - maybe we can even get a nice
interface built into FIFE.

`convertToOverlays` was moved from `initialize()` to `add_overlay()`:
Initialize would just be able to call this for the current action (e.g.
`idle`) but there is no guarantee that add_overlay is called for `idle`
only. So now before adding overlays, we check whether that is actually
possible and if not make it possible.