ITS tool: Tighten translation rule for credits, disallow everything
This has two main reasons: Most of the content in our current credits
translation is language names, and translating those doesn't feel like
something worth bothering our translators with to me personally.

Additionally, credits.xml is automatically generated from AUTHORS.md by
our helper script generate_credits. This means that if we really wanted
content of credits.xml to be translated, said script needed adaption to
include unique widget names for the text we do want to translate, while
not including anything like that for the text we do not.

All in all I do not believe adding a framework in generate_credits to
deal with this situation is a great investment of both our resources and
the ones of our translators.