mp-server: Reword error message (Kilian)
The old message had several flaws and did not fit well with our general
tone exposed towards players. Other issues were translations of the
string, up to the point where we noticed the English version would be
deemed offensive in other languages.
Not including a link to tickets tracking the work happening to improve
the situation - feel free to add that (possibly not i18n'd) if you like.