Handle overlays that become nonexistent for some reason
This usually happens if, after a tier upgrade of some building that has
this component, either a new resource can be produced that is not listed
in the inventory overlay component definition, or new action sets are
unlocked in that tier that have no such overlays defined yet.
A fix muting the warning always is to define empty structures, like

    as_i_have_no_overlays:
      RES.INPUT:
        - [0, null]
      RES.OUTNPUT:
        - [0, null]

Ideas for some more readable approach very welcome, as the above
essentially is a very verbose noop. InventoryOverlayComponent defines a
currently unused `properties` attribute that may come in handy here.