Allow scenario savegames to load again
This was (incredibly) broken since April, more precisely as consequence
of the changes in c5fb1640812e549dee60e7336edf07a16adb7ebf.

Before, the last item we parsed was a list and not a dictionary.
After the changes, the closing } of `metadata` was stripped together
with the closing } of the entire dump, as observed here:

>>> 'hi222'.rstrip('2')
'hi'
>>> 'hi222'.rsplit('2', 1)
['hi22', '']

Now try and only really remove the last brace instead of removing two.

Fixes #2164.