Introduce content/packages/
This is supposed to hold various data related to our packaging process.
As a start, copy the .xpm icon into there. Copy (as opposed to move)
because our tutorial scenarios, unlikely enough, use the old path still.
When saving a scenario game, savegames store a dump of the YAML file
used - so this is a case for savegame upgrade in any case.