Display hotkey glyphs instead of Fife key names in dynamic tooltips
This helps achieve less ugly and generally shorter text, which is good
to have in a tooltip.
When some action is not assigned at all, we display "Unassigned" which
is very long, but in that case clarity is rather important so that
nobody looks for keys with names similar to whatever abbreviation of it.