End stick-poking
This was introduced to mimic animation overlays, but those (in the
tested use cases) are easier... due to not having changing actions.
Probably as soon as inventory overlays receive actual stress testing
they will suffer from much the same problem: storing data for some
action (idle, move, work, etc.) and then changing the action needs to
be clearly communicated between Fife and all involved overlays.

Because that might be horrible and a lot of work, have a TODO instead.