Balance tweaks for warehouse, storage tent, lookout
With the introduction of a meaningful settlement range, maybe we can
make both lookout and storage tent exist in a meaningful way. That
eventually will require more differentiation, but let's start with less
running costs and no tools as construction material for the lookout.
Warehouse also has a small upkeep, rent (or however we call it) for now.

Right now there is no GUI indication about running costs existing for
these buildings, we should add that as well before a release.