Improve trade route interaction
As mentioned in issue #1688, the dialog used to manage ship's trade route has
some usability problems. In this commit I address the default amount
of goods loaded/unloaded. When user selects a new resource for a ship's slot,
the default action is to load 0t, which makes no sense and is hardly the
intention, so user always needs to update the value.

By setting the default to ships full capacity, the load order makes sense with
no further interaction. Of course user can still change the value.

The fix applied in this commit alters the behaviour only when a new resource is
set for an empty slot. In such cases, full ship's capacity is used. On the other
hand, when the slot was previously occupied by another resource, the value might
have been set manually by the user and he may only want to change resource's
type, so no alteration of the value takes place.

This partially fixes #1688.