Be smart about (un)loading in trade route's setup
As mentioned in issue #1688, the dialog used to manage ship's trade route has
some usability problems. Here I deal with the fact that whenever a resource is
selected, it defaults to loading that resource. For most basic routes resource
is loaded in one place and unloaded in another. This commit modifies the dialog
so it defaults to such setup.

Now, whenever a resource is picked, the whole route is checked to see if the
same resource is loaded or unloaded before in the route. If it is loaded, the
action defaults to unloading it. And vice-versa.

This partially fixes #1688.