Add unload-everything action to trade routes
As suggested in issue #1688, it is desired to have a way of unloading everything
from the ship. This commit implements that.

On the trade route configuration dialog, I've changed how entries (waypoints)
are displayed and added a button that turns the entry into an unload-all
position. I mark such entries by displaying a placeholder resource (a wooden
container box that could contain anything) with the unload icon. The entry can
be reverted back to a normal one by clicking on it.

Unload all entries are passed to world.traderoute.TradeRoute and handled
specially there. Upon arrival, a transfer is created based on everything present
on the ship. Wait-for-unload option is honored normally.

Save game format was updated so the new entries can be saved/loaded.