Allow different collector unit graphics per tier
When a new tier is reached, a message is broadcasted to all buildings,
which lets them upgrade their level and with that their graphics.
Collectors are units, that belong to a building. This commit lets a
building tell its collectors, that its level has changed and their level
should be changed, too. Other than that, `BuildingCollector`s now also
have a `level_upgrade()` method to update their current `_action_set_id`
to a random one of the new level.