Improve the pirate ships
One flag (with a pirate logo) of the pirate ships showed in the wrong
direction. This flag has been removed and the logo is now placed on the
main sail of the ship. So it can be seen easier.

In addition to that, both pirate ships have now a movement animation and
shadows (before just the action set 'as_pitate1').