Fit 4 related buildings into a line
Instead of arranging the related buildings by settler level (TIER), as
many buildings as possible are put in the same line. Currently, the
limit is 4.

This is necessary, because the farm had more than 4 buildings in the
same TIER and there seems to be no automatic line-break possible with
our GUI.

The grouping by TIER got removed, because there was no visible indicator
that any grouping took place and with more than 4 buildings in the same
TIER, there could be multiple interpretations of the seen building's
TIERs.