Run GUI tests in travis-ci
We are currently using manually built packages of fife for not having to
rebuild fife every time. But for some reason, travis-ci doesn't find the
.so files, when they have the (standard?) ".x86_64-linux-gnu" included
in the name. Take a look at
  `python -c "import imp; print imp.get_suffixes()"`

We are using a special resolution, because some tests didn't work at
first when using "/etc/init.d/xvfb start" but worked locally. Using a
(higher?) resolution fixed this.

Thanks to MarkusHackspacher for starting the work on this.