Fix two gui tests
The boatbuilder test asserted that some elements did not exist in the
GUI, however that must have changed. They exist but aren't visible, so
change the assertion accordingly.

The names of the buttons in the build menu changed. The tutorial test
couldn't place a building because a stone deposit was in its way.