Properly clean up when ShipOverviewTab is hidden
The base class OverviewTab and TabInterface take care of
creating/destroying gui elements, however ShipOverviewTab uses a
RouteConfig window in addition to its own widget.

The problem here is, that once the tab is hidden (and its own widget is
freed), the RouteConfig stayed available in pychan's internals. That in
turn became a problem for the test_traderoute gui tests, because
suddenly there were multiple widgets named 'configure_route' in the
system.