AI: Check for buildability when choosing a warehouse location
When the AI wants to found a new settlement, it did not check if a
warehouse could actually be built on the chosen location. The result
was, that the game crashed when the ship arrived to build the warehouse.
We are now filterting the available spots on the island for buildability
using `check_build()` of the warehouse class.