Fix production overview not showing produced resources
I'm fairly certain the issue was introduced in 14121bcb56f50f01231,
which was a fix for #1848 [1].

The issue is that both `ShipProducer` and `GroundUnitProducer` inherit
from `Producer`, but keep the same component name.
`ComponentHolder.has_component` checks existance of a component by
looking for the name. Which means that a condition like

  has_component(Producer) and not has_component(ShipProducer)

is always false, since both components have the same name "producer".

Fix this by specifying on the specific producer components whether or
not they actually produce a resource, and use that to decide whether it
should show up in the production overview or not.

Fixes #2369

[1] https://github.com/unknown-horizons/unknown-horizons/issues/1848