Fix crash when pausing a game started with speed zero
The issue was that `paused_ticks_per_second` is only ever defined once a
game is paused the first time. Starting a game with speed 0 means the
game is already paused and an attempt is made to unpause it, accessing
paused_ticks_per_second.

This seems to not break anything else - I wasn't sure about the initial
value at first, but it seems to be fine. We definitly need to define all
instance attributes in `__init__`!

Fixes #2419