Fix traderoute gui test
The GUI of the RouteConfig widget is initialized lazily because of
multiplayer games.

Since the test founds a settlement (selects a ship) and immediately (in
the same tick) tries to show the routeconfig widget, which will fail.

Hopefully fix this by waiting more ticks than are used in the delay in
`RouteConfig.__init__`.