Fix bug in selection group initialization
There is a difference between `[set() for _ in range(10)]` and
`[set()]*10`. The former creates a list with ten unique sets, the latter
a list with 10 references *to the same set*. This was introduced in
904d80.

What this means is that once you change one selection group, you change
all of them.

The selection_groups test broke suddenly, because I changed it to use a
fixture instead of a fresh dev map (reason was that the random map
algorithm changed and blocked the test from placing a warehouse). When
starting a new game, the player's ship is automatically assigned to
group 1. Haven't found out where, but that's how it is.

When you save and load the game, it iterates over all groups in the
savegame and loads each of them. But accidently setting all of them to
the same.

That's why the GUI test was unable to deselect all instances at the end,
because selection group 0 was the same as all the others, it contained
the player's ship.