Redraw messagewidget when message is removed
Everytime `self.active_messages` is modified, the widget is redrawn with
`self.draw_widget`. This happens when messages get added or are removed
in `self.tick`.

Consequently this should happen when a message is removed as well.
Thinking about it, the previous behaviour was buggy. A message was
removed internally, but the widget wasn't refreshed, therefore it still
showed up on the screen. This doesn't happen anymore now.

Fixes #2273