Center wheel button on production tabs correctly
`center_y`, opposite to what I expected, wasn't really the center of the
middle arrow. The issue is that some at the start, the center is
calculated by multiplying the resource widget's height (half of it) by
the difference between the number of input and output resources.

    height_diff_y = max(1, output_amount, input_amount) - 1
    center_y = (self.ICON_HEIGHT // 2) * height_diff_y

However, it subtracted one, which means in the case of 1 input and 1
output, center_y would be zero - which isn't what we want. Changing that
code results in bugs to the arrows, and I'm not going into that code at
this point.

Instead, fix the return value of `_center_production_line` but adding
that missing `self.ICON_HEIGHT // 2`. With that fix, we can center the
toggle button intuitively: take center_y, and subtract half of the
icon's height.

Fixes: #1901