Lazy load RouteConfig gui only when necessary
When creating a new RouteConfig instance, the code would check whether
it needed to call the CreateRoute command and only then would it
initialize the gui.

However this is only necessary, when the instance has no preconfigured
route to begin with. But it does when loading a savegame. In this case,
we can initialize the gui right away.